containers/tesseract/requirements.txt
fastapi
uvicorn
pytesseract
python-multipart
Pillow
opencv-python-headless
numpy
